requirements.txt
Django==1.7
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.5.3
Pygments==1.6
Sphinx==1.2.3
dj-database-url==0.3.0
dj-static==0.0.6
django-appconf==0.6
django-extensions==1.4.0
django-floppyforms==1.1
django-imagekit==3.2.2
django-localflavor==1.0
django-model-utils==2.2
django-pipeline==1.3.25
django-select-multiple-field==0.3.0
django-toolbelt==0.0.1
django-widget-tweaks==1.3
docutils==0.12
easypost==2.0.11
factory-boy==2.4.1
futures==2.1.6
fuzzywuzzy==0.3.1
gunicorn==19.1.1
invoke==0.9.0
pilkit==1.1.9
psycopg2==2.5.4
requests==2.4.3
six==1.7.3
sphinx-rtd-theme==0.1.6
static3==0.5.1
stripe==1.19.0
